+++ radial_menu_backend/CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog for package radial_menu_backend
 0.4.0 (2020-XX-XX)
 ------------------
 * Support null items
+* Use std's shared_ptr instead of boost's
 
 0.3.4 (2020-08-14)
 ------------------